[PATCH 1/1] packages/rtl8723bs: The rtl8723bs is now a Linux staging driver since kernel 4.12.
The old git commit is used for kernel < 4.12 with patches applied, but
for kernel >= 4.12, the newest commit featuring only binary firmwares
is fetched and only the firmwares are installed into the target
directory.

Getting the kernel version this early in the post-patch hook is only
possible from the configuration, as the kernel is not yet configured
and cannot be probed for the actual version.

Signed-off-by: Michel Stempin <michel.stempin@wanadoo.fr>